Разбил код на модули, настроил вебпак
